backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autoflake==2.3.1
black==24.4.2
click==8.1.7
decorator==5.1.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
executing==2.0.1
flake8==7.1.0
inflection==0.5.1
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
Pygments==2.18.0
pyright==1.1.370
python-dotenv==1.0.1
pytils==0.4.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
wcwidth==0.2.13
